refactor and add more rpc methods
